refactor: filters and display filters folder structure
