--- web/components/issue-layouts/index.ts
@@ -1,4 +1,0 @@
-export * from "./display-filters";
-export * from "./filters";
-export * from "./helpers";
-export * from "./layout-selection";
