+++ web/components/issues/issue-layouts/header/display-filters/sub-group-by.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { observer } from "mobx-react-lite";
 
 // components
-import { FilterHeader, FilterOption } from "components/issue-layouts";
+import { FilterHeader, FilterOption } from "components/issues";
 // types
 import { TIssueGroupByOptions } from "types";
 // constants
