+++ web/components/issues/issue-layouts/index.ts
@@ -1,2 +1,3 @@
 export * from "./calendar";
+export * from "./header";
 export * from "./kanban";
